refactor: rename tabs-plugin
