+++ example/app/data/DemoDataSource/DemoPackage/recipe_links/tabs_as_sidebar_blueprint.json
@@ -2,11 +2,11 @@
   "type": "CORE:RecipeLink",
   "_blueprintPath_": "blueprints/TabsAsSidebarBlueprint",
   "initialUiRecipe": {
-    "name": "Tabs",
+    "name": "AttributeSelector",
     "type": "CORE:UiRecipe",
-    "plugin": "tabs",
+    "plugin": "@development-framework/dm-core-plugins/attribute-selector",
     "config": {
-      "type": "blueprints/Plugins/TabsPluginConfig",
+      "type": "blueprints/Plugins/AttributeSelectorPluginConfig",
       "childTabsOnRender": true,
       "asSidebar": true,
       "homeRecipe": "home"
