+++ packages/dm-core-plugins/src/attribute-selector/Content.tsx
@@ -6,7 +6,7 @@ import {
   UIRecipesSelector,
 } from '@development-framework/dm-core'
 import { useTabContext } from './TabsContext'
-import { TChildTab } from './TabsPlugin'
+import { TChildTab } from './AttributeSelectorPlugin'
 
 const HidableWrapper = styled.div<any>`
   display: ${(props: { hidden: boolean }) => (props.hidden && 'none') || 'flex'}
